# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Jaga Asset Management API",
        "contact": {}
    },
    "paths": {
...
        "/v1/maintenance-records": {
            "get": {
                "description": "Retrieve maintenance records with optional filters and pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaintenanceRecords"
                ],
                "summary": "Get a list of maintenance records",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetMaintenanceRecordsResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a maintenance record for an asset",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaintenanceRecords"
                ],
                "summary": "Create a new maintenance record",
                "parameters": [
                    {
                        "description": "Create Maintenance Record",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateMaintenanceRecordRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateMaintenanceRecordResponse"
                        }
                    }
                }
            }
        },
        "/v1/maintenance-records/{id}": {
            "get": {
                "description": "Retrieve a maintenance record by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaintenanceRecords"
                ],
                "summary": "Get a maintenance record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Maintenance Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetMaintenanceRecordByIDResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update the details of an existing maintenance record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaintenanceRecords"
                ],
                "summary": "Update an existing maintenance record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Maintenance Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Maintenance Record",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateMaintenanceRecordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateMaintenanceRecordResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete the maintenance record identified by the ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaintenanceRecords"
                ],
                "summary": "Delete a maintenance record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Maintenance Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.DeleteMaintenanceRecordResponse"
                        }
                    }
                }
            }
        },
        "/v1/maintenance-records/{id}/status": {
            "put": {
                "description": "Update the status of a specific maintenance record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaintenanceRecords"
                ],
                "summary": "Update the status of a maintenance record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Maintenance Record ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Maintenance Record Status",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateMaintenanceRecordStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateMaintenanceRecordResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateMaintenanceRecordRequest": {
            "type": "object",
            "required": [
                "asset_id",
                "description",
                "maintenance_date",
                "status"
            ],
            "properties": {
                "asset_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 5
                },
                "maintenance_date": {
                    "type": "string"
                },
                "performed_by": {
                    "type": "string"
                },
                "schedule_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "in_progress",
                        "on_hold",
                        "finished",
                        "failed",
                        "cancelled"
                    ]
                }
            }
        },
        "dto.CreateMaintenanceRecordResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.DeleteMaintenanceRecordResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetMaintenanceRecordByIDResponse": {
            "type": "object",
            "properties": {
                "maintenance_record": {
                    "$ref": "#/definitions/dto.MaintenanceRecordDTO"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.GetMaintenanceRecordsResponse": {
            "type": "object",
            "properties": {
                "items_per_page": {
                    "type": "integer"
                },
                "maintenance_records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.MaintenanceRecordDTO"
                    }
                },
                "message": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "dto.MaintenanceRecordDTO": {
            "type": "object",
            "properties": {
                "asset_id": {
                    "type": "string"
                },
                "asset_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "maintenance_date": {
                    "type": "string"
                },
                "performed_by_user_id": {
                    "type": "string"
                },
                "performed_by_user_name": {
                    "type": "string"
                },
                "schedule_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateMaintenanceRecordRequest": {
            "type": "object",
            "properties": {
                "asset_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 5
                },
                "maintenance_date": {
                    "type": "string"
                },
                "performed_by": {
                    "type": "string"
                },
                "schedule_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "in_progress",
                        "on_hold",
                        "finished",
                        "failed",
                        "cancelled"
                    ]
                }
            }
        },
        "dto.UpdateMaintenanceRecordResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateMaintenanceRecordStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "in_progress",
                        "on_hold",
                        "finished",
                        "failed",
                        "cancelled"
                    ]
                }
            }
        },
...
    }
}